+++ app/Parse_mods/offence_stats/Attack_Critical_Chance.php
@@ -15,7 +15,7 @@ class Attack_Critical_Chance extends Stat
     {
         $this->setType($type);
         $this->name ='% Critical Chance for Attacks';
-        $modValue = filter_var($mod, FILTER_SANITIZE_NUMBER_INT);
+        $modValue = (int) filter_var($mod, FILTER_SANITIZE_NUMBER_INT);
         $modName = preg_replace('/\d+/u', '#', $mod);
 
         // Combined Critical Chance From different Weapon Types
